backend/requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
django-filter>=24.0
psycopg2-binary>=2.9,<3.0
dj-database-url>=2.1,<3.0
python-decouple>=3.8,<4.0
gunicorn>=22.0,<23.0
